requirements.txt
scrapy
pandas
newspaper3k
nest_asyncio
requests
beautifulsoup4
openpyxl
transformers
ollama
tqdm
